refactor: run boost pool setup in parallel
