+++ bouncer/commands/setup_boost_pools.ts
@@ -4,76 +4,15 @@
 // This command takes no arguments.
 // This command will create 3 tiers of boost pools for every asset. Tiers: 5, 10 and 30 bps.
 
-import { InternalAsset as Asset, Chain } from '@chainflip/cli/.';
-import {
-  ingressEgressPalletForChain,
-  getAssetsForChain,
-  getChainflipApi,
-  observeEvent,
-} from '../shared/utils';
-import { submitGovernanceExtrinsic } from '../shared/cf_governance';
-
-type BoostPoolId = {
-  asset: Asset;
-  tier: number;
-};
-
-// These are the tiers of boost pools that will be created for each asset
-const boostPoolTiers = [5, 10, 30];
-const chains = ['Ethereum', 'Polkadot', 'Bitcoin', 'Arbitrum'];
+import { setupBoostPools } from '../shared/setup_boost_pools';
+import { runWithTimeout } from '../shared/utils';
 
 async function main(): Promise<void> {
-  console.log('=== Creating Boost Pools ===');
-  const chainflip = await getChainflipApi();
-  const observeBoostPoolEvents = [];
-
-  for (const c of chains) {
-    const chain = c as Chain;
-    console.log(`Creating boost pools for all ${chain} assets`);
-    const assets = await getAssetsForChain(chain);
-    const newPools: BoostPoolId[] = [];
-
-    for (const asset of assets) {
-      for (const tier of boostPoolTiers) {
-        if (tier <= 0) {
-          throw new Error(`Invalid tier value: ${tier}`);
-        }
-        newPools.push({ asset, tier });
-
-        const observeBoostPoolCreated = observeEvent(
-          `${chain.toLowerCase()}IngressEgress:BoostPoolCreated`,
-          chainflip,
-          (event) =>
-            event.data.boostPool.asset === asset && event.data.boostPool.tier === tier.toString(),
-        );
-        const observeBoostPoolAlreadyExists = observeEvent(`governance:FailedExecution`, chainflip);
-
-        observeBoostPoolEvents.push(
-          Promise.race([observeBoostPoolCreated, observeBoostPoolAlreadyExists]),
-        );
-      }
-    }
-
-    const ingressEgressPallet = await ingressEgressPalletForChain(chain);
-    submitGovernanceExtrinsic(ingressEgressPallet.createBoostPools(newPools));
-  }
-
-  const boostPoolEvents = await Promise.all(observeBoostPoolEvents);
-  for (const event of boostPoolEvents) {
-    if (event.name.method !== 'BoostPoolCreated') {
-      // TODO: decode error here
-      throw new Error(`Failed to create boost pool: ${JSON.stringify(event)}`);
-    }
-    console.log(
-      `Boost pools created for ${event.data.boostPool.asset} at ${event.data.boostPool.tier} bps`,
-    );
-  }
-
-  console.log('=== Boost Pools Setup completed ===');
+  await setupBoostPools();
   process.exit(0);
 }
 
-main().catch((error) => {
+runWithTimeout(main(), 60000).catch((error) => {
   console.error(error);
   process.exit(-1);
 });
