+++ bouncer/commands/setup_concurrent.ts
@@ -2,14 +2,15 @@
 // INSTRUCTIONS
 //
 // This command takes no arguments.
-// It will setup pools and zero to infinity range orders for all currencies
-// For example: ./commands/setup_swaps.ts
-
+// It will setup pools, zero to infinity range orders and boost pools for all currencies.
+// For example: ./commands/setup_concurrent.ts
+// Setup_vaults.ts must be ran first.
+import { setupBoostPools } from '../shared/setup_boost_pools';
 import { setupSwaps } from '../shared/setup_swaps';
 import { runWithTimeout } from '../shared/utils';
 
 async function main(): Promise<void> {
-  await setupSwaps();
+  await Promise.all([setupSwaps(), setupBoostPools()]);
   process.exit(0);
 }
