+++ bouncer/setup_for_test.sh
@@ -1,7 +1,6 @@
 set -e
 ./commands/setup_vaults.ts
-./commands/setup_swaps.ts
-./commands/setup_boost_pools.ts
+./commands/setup_concurrent.ts
 
 echo "✅ Ready for test to run"
